requirements.txt
tensorflow==2.8.0
flask==2.0.1
numpy==1.19.5
pillow==8.3.1
scikit-learn==0.24.2
joblib==1.0.1
pandas==1.3.0 